Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic>=2.7.4
pydantic-settings
python-jose[cryptography]
bcrypt==3.1.7 
passlib
python-multipart
python-dotenv
alembic
pydantic[email]
requests
PyPDF2
pandas
langchain
langchain_huggingface
langchain_groq
chromadb>=0.4.0
# python-magic
python-magic-bin
sentence-transformers
beautifulsoup4
validators
